doc: add a page about tombstone GC
This commit adds a page dedicated to tombstone
garbage collection (GC) and its modes.
It extends the existing documentation on
the Data Definition page, provides an overview,
and makes the information easily accessible.

The link to this page is added in appropriate places across the documentation.

Fixes https://github.com/scylladb/scylladb/issues/14000

Refs https://github.com/scylladb/scylladb/issues/14130