docs: nodetool compact: remove unused and unrecognized options
The doc page for nodetool compact lists a bunch of JMX related options
that are just distracting. These are common to all nodetool commands, no
need to list them just on the page for nodetool compact.
Also remove the --partition option, which is not even recognized by the
current nodetool.