test: Verify server exit code during graceful process shutdown.
Currently, it's possible for a test to pass even if the server crashes
during a graceful shutdown. Additionally, the server may crash in the
middle of a test, resulting in a test failure with an inaccurate
description.  This commit updates the test framework to monitor the
server's return code and throw an exception in the event of an abnormal
server shutdown.

Fixes scylladb/scylla#15365

Closes scylladb/scylladb#15660