node_ops: add tasks covering raft join request
Raft join request is inserted in two different ways depending on whether
a node is the first one to join a token ring. Thus, two different task
are added:
- raft_first_bootstrap_task_impl, which covers the bootstrap of the first
  node in a token ring, when joining node submits the requests.
- raft_joining_entry_task_impl, which covers the bootstrap of the following
  nodes in a token ring, when join request is submitted by an existing node
  during the handshake. The task is created on the submitting node.

In the future, we will be able to link raft_joining_entry_task_impl
with the tasks initiating the bootstrap (join_token_ring_task_impl), and
to recognise and manage a task which covers the bootstrap/replace
streaming using tasks introduced in this commit.