build: set an older version for cxxbridge that works in the frozen toolchain
In this branch(5.1) the most recent available rustc version is 1.60,
despite that, the 'cargo install' command tries to install the most
recent version of a package by default, which may rely on newer rustc
versions. This patch specifies the version of the cxxbridge-cmd package
to one that works with rustc 1.60.

Closes scylladb/scylladb#15812

[avi: regenerated frozen toolchain]