test.py: move to a gracefull temination of nodes on teardown
This change move existing suits which create cluster through the
testing infra to be stopped and uninstalled gracefully.
The motivation, besides the obvious advantage of testing our stop
sequence is that it will pave the way for applying code coverage support
to all tests (not only standalone unit and boost test executables).

testing:
	Ran all tests 10 times in a row in dev mode.
	Ran all tests once in release mode
	Ran all tests once in debug mode

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>