build: create outdir when outdir is explictly used
actually we've created outdir when using it as the parent directory
of `tempfile.tempdir`, but there are many places where we use
`tempfile.tempdir` for, for instance, testing the compiler flags,
and these tests will be removed once we migrate to CMake, so they
do not really matter when reviewing the change which migrates to
CMake.

the point of this change is to help the review understand the major
changes performed by the migration.

Refs #15379
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>