raft: allow to predeclare group0 ID in group registry
The group0 ID is immediately known after discovery. Allow setting it in
raft_group_registry even before group0 server is started.

Declare the group0 ID after discovery and before join node request RPC
is sent to the cluster. The node will not be added to group0 config
until join node request is sent, so this allows us to properly classify
a raft RPC as group0/non-group0 (unless the node was added during the
previous attempt to boot, but re-attempting the join procedure is not
allowed and will fail anyway).