build: stop passing --date-stamp option to SCYLLA-VERSION-GEN
`SCYLLA-VERSION-GEN` does not reference the date-stamp specified with
this option at all. instead, it always overwrites it with the the output
of `date --utc +%Y%m%d`, if we are going to reference this value.

so, in this change, instead of pretending that we support this option,
let's just drop it. because this option is still used by our CI
workflow, let's keep it around before this option is removed there as well.
we will drop this option in `SCYLLA-VERSION-GEN` in an other commit.

for those who are interested, this option stopped working
since 839d8f40, which practically traded --date-stamp option for
the env variable of `SCYLLA_RELEASE`.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>