SCYLLA-VERSION-GEN: drop --date-stamp option
the argument of this option is assigned to `DATE` variable, which is
always overwritten with the the output of `date --utc +%Y%m%d`, if
we are going to reference this value.

so, in this change, instead of pretending that we support this option,
let's just drop it in this script. because this script is always called
by `configure.py`, and we've stopped passing this option in
`configure.py`, it's safe to drop this option here.

for those who are interested, this option stopped working since
839d8f40, which practically traded --date-stamp option for the env
variable of `SCYLLA_RELEASE`.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>