raft: topology: only pull topology snapshot in topology-on-raft mode
Currently, during group0 snapshot transfer, the node pulling
the snapshot will send the `raft_pull_topology_snapshot` verb even if
the cluster is not in topology-on-raft mode. The RPC handler returns an
empty snapshot in that case. However, using the verb outside topology on
raft causes problems:

- It can cause issues during rolling upgrade as the snapshot transfer
  will keep failing on the upgraded nodes until the leader node is
  upgraded,
- Topology changes on raft are still experimental, and using the RPC
  outside experimental mode will prevent us from doing breaking changes
  to it.

Solve the issue by passing the "topology changes on raft enabled" flag
to group0_state_machine and send the RPC only in topology on raft mode.