cache_flat_mutation_reader: fix a logic bug in ensure_population_lower_bound() with reverse reads
`_last_row` is in table schema, while `cur.position()` is in query schema
(which is either equal to table schema, or its reverse).

Thus, the comparison affected by this patch doesn't work as intended.
In reverse reads, the check will pass even if `_last_row` has the same key,
but opposite bound weight to `cur`, which will lead to the dummy being inserted
at the wrong position, which can e.g. wrongly extend a range tombstone.

Fix that.